fix: moves slug func to utils, edits save method for Category model
